backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
google-cloud-storage>=2.14.0
google-cloud-firestore>=2.14.0
redis>=5.0.0
python-dotenv>=1.0.0
qrcode>=7.4.2
Pillow>=10.2.0
loguru>=0.7.2
